Cleanup `data_view3`
